adding owner_id in pets

+++ src/main/resources/db/migration/V202408091500__create_owners_table.sql
@@ -2,7 +2,7 @@ create table owners
 (
     id serial primary key NOT NULL,
     name varchar NOT NULL,
-    company_id int NOT NULL,
+    company_id BIGINT NOT NULL,
     employee_id varchar NOT NULL
 )
 CREATE INDEX idx_owners_company_id_employee_id ON owners (company_id, employee_id);
